refactor: refactor a lot
BREAKING CHANGE: this is a breaking change
