+++ src/so_vits_svc_fork/inference/inference_main.py
@@ -10,7 +10,7 @@ import soundfile
 import torch
 from cm_time import timer
 
-from .inference.infer_tool import RealtimeVC, RealtimeVC2, Svc
+from so_vits_svc_fork.inference.infer_tool import RealtimeVC, RealtimeVC2, Svc
 
 LOG = getLogger(__name__)
