+++ src/so_vits_svc_fork/modules/decoders/hifigan/__init__.py
@@ -1,0 +1,3 @@
+from ._models import Generator
+
+__all__ = ["Generator"]
