+++ src/so_vits_svc_fork/modules/decoders/hifigan/_models.py
@@ -7,7 +7,7 @@ import torch.nn.functional as F
 from torch.nn import AvgPool1d, Conv1d, Conv2d, ConvTranspose1d
 from torch.nn.utils import remove_weight_norm, spectral_norm, weight_norm
 
-from .utils import get_padding, init_weights
+from ._utils import get_padding, init_weights
 
 LOG = getLogger(__name__)
