+++ src/so_vits_svc_fork/modules/onnx/model_onnx.py
@@ -1,22 +1,14 @@
-from logging import getLogger
-from typing import Any, Literal, Sequence
-
 import torch
 from torch import nn
 from torch.nn import Conv1d, Conv2d
 from torch.nn import functional as F
 from torch.nn.utils import spectral_norm, weight_norm
 
-import so_vits_svc_fork.modules.attentions as attentions
-import so_vits_svc_fork.modules.commons as commons
-import so_vits_svc_fork.modules.modules as modules
-
-from . import utils
-from .modules.commons import get_padding
-from .utils import f0_to_coarse
-from .vdecoder.hifigan.models import Generator
-
-LOG = getLogger(__name__)
+import so_vits_svc_fork.f0
+from so_vits_svc_fork.f0 import f0_to_coarse
+from so_vits_svc_fork.modules import attentions, commons, modules
+from so_vits_svc_fork.modules.commons import get_padding
+from so_vits_svc_fork.modules.decoders.hifigan import Generator
 
 
 class ResidualCouplingBlock(nn.Module):
@@ -132,13 +124,12 @@ class TextEncoder(nn.Module):
             hidden_channels, filter_channels, n_heads, n_layers, kernel_size, p_dropout
         )
 
-    def forward(self, x, x_mask, f0=None, noice_scale=1):
+    def forward(self, x, x_mask, f0=None, z=None):
         x = x + self.f0_emb(f0).transpose(1, 2)
         x = self.enc_(x * x_mask, x_mask)
         stats = self.proj(x) * x_mask
         m, logs = torch.split(stats, self.out_channels, dim=1)
-        z = (m + torch.randn_like(m) * torch.exp(logs) * noice_scale) * x_mask
-
+        z = (m + z * torch.exp(logs)) * x_mask
         return z, m, logs, x_mask
 
 
@@ -251,85 +242,6 @@ class DiscriminatorS(torch.nn.Module):
         return x, fmap
 
 
-class MultiPeriodDiscriminator(torch.nn.Module):
-    def __init__(self, use_spectral_norm=False):
-        super().__init__()
-        periods = [2, 3, 5, 7, 11]
-
-        discs = [DiscriminatorS(use_spectral_norm=use_spectral_norm)]
-        discs = discs + [
-            DiscriminatorP(i, use_spectral_norm=use_spectral_norm) for i in periods
-        ]
-        self.discriminators = nn.ModuleList(discs)
-
-    def forward(self, y, y_hat):
-        y_d_rs = []
-        y_d_gs = []
-        fmap_rs = []
-        fmap_gs = []
-        for i, d in enumerate(self.discriminators):
-            y_d_r, fmap_r = d(y)
-            y_d_g, fmap_g = d(y_hat)
-            y_d_rs.append(y_d_r)
-            y_d_gs.append(y_d_g)
-            fmap_rs.append(fmap_r)
-            fmap_gs.append(fmap_g)
-
-        return y_d_rs, y_d_gs, fmap_rs, fmap_gs
-
-
-class SpeakerEncoder(torch.nn.Module):
-    def __init__(
-        self,
-        mel_n_channels=80,
-        model_num_layers=3,
-        model_hidden_size=256,
-        model_embedding_size=256,
-    ):
-        super().__init__()
-        self.lstm = nn.LSTM(
-            mel_n_channels, model_hidden_size, model_num_layers, batch_first=True
-        )
-        self.linear = nn.Linear(model_hidden_size, model_embedding_size)
-        self.relu = nn.ReLU()
-
-    def forward(self, mels):
-        self.lstm.flatten_parameters()
-        _, (hidden, _) = self.lstm(mels)
-        embeds_raw = self.relu(self.linear(hidden[-1]))
-        return embeds_raw / torch.norm(embeds_raw, dim=1, keepdim=True)
-
-    def compute_partial_slices(self, total_frames, partial_frames, partial_hop):
-        mel_slices = []
-        for i in range(0, total_frames - partial_frames, partial_hop):
-            mel_range = torch.arange(i, i + partial_frames)
-            mel_slices.append(mel_range)
-
-        return mel_slices
-
-    def embed_utterance(self, mel, partial_frames=128, partial_hop=64):
-        mel_len = mel.size(1)
-        last_mel = mel[:, -partial_frames:]
-
-        if mel_len > partial_frames:
-            mel_slices = self.compute_partial_slices(
-                mel_len, partial_frames, partial_hop
-            )
-            mels = list(mel[:, s] for s in mel_slices)
-            mels.append(last_mel)
-            mels = torch.stack(tuple(mels), 0).squeeze(1)
-
-            with torch.no_grad():
-                partial_embeds = self(mels)
-            embed = torch.mean(partial_embeds, axis=0).unsqueeze(0)
-            # embed = embed / torch.linalg.norm(embed, 2)
-        else:
-            with torch.no_grad():
-                embed = self(last_mel)
-
-        return embed
-
-
 class F0Decoder(nn.Module):
     def __init__(
         self,
@@ -378,30 +290,26 @@ class SynthesizerTrn(nn.Module):
 
     def __init__(
         self,
-        spec_channels: int,
-        segment_size: int,
-        inter_channels: int,
-        hidden_channels: int,
-        filter_channels: int,
-        n_heads: int,
-        n_layers: int,
-        kernel_size: int,
-        p_dropout: int,
-        resblock: str,
-        resblock_kernel_sizes: Sequence[int],
-        resblock_dilation_sizes: Sequence[Sequence[int]],
-        upsample_rates: Sequence[int],
-        upsample_initial_channel: int,
-        upsample_kernel_sizes: Sequence[int],
-        gin_channels: int,
-        ssl_dim: int,
-        n_speakers: int,
-        sampling_rate: int = 44100,
-        type_: Literal["hifi-gan", "istft", "ms-istft", "mb-istft"] = "hifi-gan",
-        gen_istft_n_fft: int = 16,
-        gen_istft_hop_size: int = 4,
-        subbands: int = 8,
-        **kwargs: Any,
+        spec_channels,
+        segment_size,
+        inter_channels,
+        hidden_channels,
+        filter_channels,
+        n_heads,
+        n_layers,
+        kernel_size,
+        p_dropout,
+        resblock,
+        resblock_kernel_sizes,
+        resblock_dilation_sizes,
+        upsample_rates,
+        upsample_initial_channel,
+        upsample_kernel_sizes,
+        gin_channels,
+        ssl_dim,
+        n_speakers,
+        sampling_rate=44100,
+        **kwargs
     ):
         super().__init__()
         self.spec_channels = spec_channels
@@ -434,54 +342,18 @@ class SynthesizerTrn(nn.Module):
             kernel_size=kernel_size,
             p_dropout=p_dropout,
         )
-
-        LOG.info(f"Decoder type: {type_}")
-        if type_ == "hifi-gan":
-            hps = {
-                "sampling_rate": sampling_rate,
-                "inter_channels": inter_channels,
-                "resblock": resblock,
-                "resblock_kernel_sizes": resblock_kernel_sizes,
-                "resblock_dilation_sizes": resblock_dilation_sizes,
-                "upsample_rates": upsample_rates,
-                "upsample_initial_channel": upsample_initial_channel,
-                "upsample_kernel_sizes": upsample_kernel_sizes,
-                "gin_channels": gin_channels,
-            }
-            self.dec = Generator(h=hps)
-            self.mb = False
-        else:
-            hps = {
-                "initial_channel": inter_channels,
-                "resblock": resblock,
-                "resblock_kernel_sizes": resblock_kernel_sizes,
-                "resblock_dilation_sizes": resblock_dilation_sizes,
-                "upsample_rates": upsample_rates,
-                "upsample_initial_channel": upsample_initial_channel,
-                "upsample_kernel_sizes": upsample_kernel_sizes,
-                "gin_channels": gin_channels,
-                "gen_istft_n_fft": gen_istft_n_fft,
-                "gen_istft_hop_size": gen_istft_hop_size,
-                "subbands": subbands,
-            }
-            from .vdecoder.mb_istft.generators import (
-                Multiband_iSTFT_Generator,
-                Multistream_iSTFT_Generator,
-                iSTFT_Generator,
-            )
-
-            # gen_istft_n_fft, gen_istft_hop_size, subbands
-            if type_ == "istft":
-                del hps["subbands"]
-                self.dec = iSTFT_Generator(**hps)
-            elif type_ == "ms-istft":
-                self.dec = Multistream_iSTFT_Generator(**hps)
-            elif type_ == "mb-istft":
-                self.dec = Multiband_iSTFT_Generator(**hps)
-            else:
-                raise ValueError(f"Unknown type: {type_}")
-            self.mb = True
-
+        hps = {
+            "sampling_rate": sampling_rate,
+            "inter_channels": inter_channels,
+            "resblock": resblock,
+            "resblock_kernel_sizes": resblock_kernel_sizes,
+            "resblock_dilation_sizes": resblock_dilation_sizes,
+            "upsample_rates": upsample_rates,
+            "upsample_initial_channel": upsample_initial_channel,
+            "upsample_kernel_sizes": upsample_kernel_sizes,
+            "gin_channels": gin_channels,
+        }
+        self.dec = Generator(h=hps)
         self.enc_q = Encoder(
             spec_channels,
             inter_channels,
@@ -505,70 +377,30 @@ class SynthesizerTrn(nn.Module):
             spk_channels=gin_channels,
         )
         self.emb_uv = nn.Embedding(2, hidden_channels)
+        self.predict_f0 = False
 
-    def forward(self, c, f0, uv, spec, g=None, c_lengths=None, spec_lengths=None):
-        g = self.emb_g(g).transpose(1, 2)
-        # ssl prenet
-        x_mask = torch.unsqueeze(commons.sequence_mask(c_lengths, c.size(2)), 1).to(
-            c.dtype
-        )
-        x = self.pre(c) * x_mask + self.emb_uv(uv.long()).transpose(1, 2)
+    def forward(self, c, f0, mel2ph, uv, noise=None, g=None):
+        decoder_inp = F.pad(c, [0, 0, 1, 0])
+        mel2ph_ = mel2ph.unsqueeze(2).repeat([1, 1, c.shape[-1]])
+        c = torch.gather(decoder_inp, 1, mel2ph_).transpose(1, 2)  # [B, T, H]
 
-        # f0 predict
-        lf0 = 2595.0 * torch.log10(1.0 + f0.unsqueeze(1) / 700.0) / 500
-        norm_lf0 = utils.normalize_f0(lf0, x_mask, uv)
-        pred_lf0 = self.f0_decoder(x, norm_lf0, x_mask, spk_emb=g)
-
-        # encoder
-        z_ptemp, m_p, logs_p, _ = self.enc_p(x, x_mask, f0=f0_to_coarse(f0))
-        z, m_q, logs_q, spec_mask = self.enc_q(spec, spec_lengths, g=g)
-
-        # flow
-        z_p = self.flow(z, spec_mask, g=g)
-        z_slice, pitch_slice, ids_slice = commons.rand_slice_segments_with_pitch(
-            z, f0, spec_lengths, self.segment_size
-        )
-
-        # nsf decoder
-        # MB-iSTFT-VITS
-        if self.dec:
-            o, o_mb = self.dec(z_slice, g=g)
-        else:
-            o = self.dec(z_slice, g=g, f0=pitch_slice)
-            o_mb = None
-        return (
-            o,
-            o_mb,
-            ids_slice,
-            spec_mask,
-            (z, z_p, m_p, logs_p, m_q, logs_q),
-            pred_lf0,
-            norm_lf0,
-            lf0,
-        )
-
-    def infer(self, c, f0, uv, g=None, noice_scale=0.35, predict_f0=False):
         c_lengths = (torch.ones(c.size(0)) * c.size(-1)).to(c.device)
+        g = g.unsqueeze(0)
         g = self.emb_g(g).transpose(1, 2)
         x_mask = torch.unsqueeze(commons.sequence_mask(c_lengths, c.size(2)), 1).to(
             c.dtype
         )
         x = self.pre(c) * x_mask + self.emb_uv(uv.long()).transpose(1, 2)
 
-        if predict_f0:
+        if self.predict_f0:
             lf0 = 2595.0 * torch.log10(1.0 + f0.unsqueeze(1) / 700.0) / 500
-            norm_lf0 = utils.normalize_f0(lf0, x_mask, uv, random_scale=False)
+            norm_lf0 = so_vits_svc_fork.f0.normalize_f0(
+                lf0, x_mask, uv, random_scale=False
+            )
             pred_lf0 = self.f0_decoder(x, norm_lf0, x_mask, spk_emb=g)
             f0 = (700 * (torch.pow(10, pred_lf0 * 500 / 2595) - 1)).squeeze(1)
 
-        z_p, m_p, logs_p, c_mask = self.enc_p(
-            x, x_mask, f0=f0_to_coarse(f0), noice_scale=noice_scale
-        )
+        z_p, m_p, logs_p, c_mask = self.enc_p(x, x_mask, f0=f0_to_coarse(f0), z=noise)
         z = self.flow(z_p, c_mask, g=g, reverse=True)
-
-        # MB-iSTFT-VITS
-        if self.mb:
-            o, o_mb = self.dec(z * c_mask, g=g)
-        else:
-            o = self.dec(z * c_mask, g=g, f0=f0)
+        o = self.dec(z * c_mask, g=g, f0=f0)
         return o
